W211: the crop is mined out, the structure band is not — +2 MATCH (6549 -> 6551, 79.8%), nd -1752
Census build/wip/W211_crop.md offered every slice two lists: List 1, the 167 first-party
functions at or under window with nd <= 60 and no floor annotation, and List 2, the at-window
structure band at nd 60-260.  Every one of the eight agents was told to work List 1 first.
0 MATCHes lost, 0 new over-window, 0 compile errors, 0 stale markers.

**The central result of this wave is a negative, and it is worth more than the two MATCHes.
List 1 is mined out.**  Eight slices worked it independently and returned, between them, ONE
match out of 167 candidates.  Everything else in that band is a b210 floor: commutative
operand orientation, call-argument setup order before a JAL, cyclic saved-register colouring,
independent adjacent memory accesses, loop-preheader constant hoisting.  Those five families
are now annotated in place at roughly sixty new sites, so no future wave needs to rediscover
them.  A low normalized diff does NOT mean a function is close to matching; it means the
remaining difference is small, and small differences are exactly what the compiler floors are.

So mid-wave every agent was redirected to List 2, and the same eight agents that had produced
one match and sixty annotations immediately produced the largest structural gains in ten waves:

  mdlFile 0034b010        nd253 -> **nd14**   template transfer from a matched sibling
  mdlFile 00349d30        nd251 -> **nd22**   same transfer
  fclCombineEffect 0041ae20  nd237 -> **nd33**   four raw data addresses corrected, then
                          caller-specific absolute u16 aliases for 007e094e / 007e0958
  mt_evtMenu 00375590     nd205 -> **nd19**   typed clamp work, correct absolute input
                          aliases, the recovered DAT_007cb088 direction branch, exact predicate
  _h_camp_data 0016c010   nd231 -> **nd53**
  y_runtime 0044a790      nd250 -> nd139  scaled-X value lifetime reconstruction
  fclShopMisc 003fd570    nd252 -> nd156  both missing FUN_003fe020 calls restored plus an
                          absolute-array alias for DAT_006af140
  mt_evtMenu 003715b0     nd155 -> nd118, mt_evtMain 00361980 nd216 -> nd189 (template
                          transfer from the matched same-file node walker: hoist the list
                          bound once, advance the link BEFORE incrementing the counter),
                          fclShopMisc 004045d0 nd229 -> 205, mdlFile 00322880 nd249 -> 201,
                          k_encount 001d9310 nd245 -> 203 and 001d96a0 nd247 -> 208,
                          k_clump 001a9500 nd246 -> 218, btlFormation 00281f20 nd258 -> 246.

The two MATCHes:
  mdlFile 0034db30   nd6 -> 0.  The loop counter increment moved into the explicit loop body
                     AHEAD of the two pointer increments.  The only List 1 finish in the wave,
                     and notably it was a loop-SHAPE correction, not a spelling.
  k_clump 001a7fc0   template transfer from a matched sibling plus deleting a persistent
                     object local.

Two negatives that close long-running questions.

**Commutative addu orientation is now a measured floor, not folklore.**  Four itfPanel and
frTagTable functions sat at nd1-nd3 blocked by the identical word — ours 0x00621821
`addu $v1,$v1,$v0`, retail 0x00431821 `addu $v1,$v0,$v1`.  Three or four MATCHes hung on one
instruction, so it earned a dedicated pass: literal operand swaps, declaration-order swaps,
pointer-versus-integer conversion placement, compound assignment in both destination
directions, distinct result locals, and explicit typed entry pointers.  None flipped it.
Worth recording plainly: the nearby MATCHED siblings obtain the desired orientation only
through their pre-existing annotated `volatile` reads.  That is a compiler-steering
`volatile` on non-hardware data, it is banned, and the agent was right to decline to copy it
and to leave those four functions at nd1.  **A match we cannot buy honestly is not a match we
want.**  003ad1b0 did improve nd3 -> nd1 on the way, from an honest `offset + base` spelling.

**The missing-call defect in the Shop slice is exhausted.**  A whole-slice retail-jal versus
R_MIPS_26 census across all 167 NONMATCHING functions found exactly one remaining genuine
deficit, 004045d0, which was then fixed.  Three earlier waves each found instances there; the
class is now closed and should not be re-censused.

Also settled: the Camp draw recipe's depth constant is 100.0f, not 50.0f — retail shows 0x42c8
at _h_camp_data 0011ddb0, and a 50.0f probe regressed it.  And btlFormation 002c2ed0 proves the
staged call-result local is sometimes GENUINE: deleting it regressed nd180 to nd321 and shrank
the object, so Battle's signature defect needs confirming per site rather than assuming.

+++ src/Battle/btlVoice.c
@@ -4299,7 +4299,7 @@ void func_002e4310(void);
 u32 func_002e4330(u32 param_1);
 s32 func_002e43a0(void);
 u32 func_002e4430(BtlAction* action, s32 skillId);
-u32 func_002e4520(int param_1,u32 param_2,u64 param_3);
+u32 func_002e4520(int param_1,u32 param_2,u32 param_3);
 u32 func_002e4720(u16 param_1);
 u32 func_002e47a0(BtlAction* action, s32 skillId);
 u32 func_002e4910(BtlAction* action, s32 skillId);
@@ -5307,7 +5307,7 @@ u32 func_002e4430(BtlAction* action, s32 skillId)
 }
 
 // FUN_002e4520 NONMATCHING
-u32 func_002e4520(int param_1,u32 param_2,u64 param_3)
+u32 func_002e4520(int param_1,u32 param_2,u32 param_3)
 
 {
   extern u32 FUN_00301ca0();
